refactor: rename enum Type to PathType
