+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -1,5 +1,6 @@
 package nextstep.favorite.application;
 
+import nextstep.auth.exception.UnAuthorizedException;
 import nextstep.favorite.application.dto.FavoriteRequest;
 import nextstep.favorite.application.dto.FavoriteResponse;
 import nextstep.favorite.domain.Favorite;
@@ -7,18 +8,16 @@ import nextstep.favorite.domain.FavoriteRepository;
 import nextstep.member.application.MemberService;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
-import nextstep.auth.exception.UnAuthorizedException;
 import nextstep.path.service.PathFinder;
 import nextstep.station.entity.Station;
 import nextstep.station.service.StationService;
-import org.springframework.data.geo.Distance;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 import static nextstep.common.constant.ErrorCode.UNAUTHORIZED_ACCESS;
-import static nextstep.common.constant.Type.DISTANCE;
+import static nextstep.common.constant.PathType.DISTANCE;
 import static nextstep.converter.FavoriteConverter.favoritesToFavoriteResponses;
 
 @Service
